Local storage arreglado + Drag&Drop arreglado

+++ DAW/CapDeDatos/database/migrations/2024_01_29_081838_create_localizaciones_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->string('nombre')->primary();
             $table->float('latitud');
             $table->float('longitud');
-            $table->float('temperatura');
+            $table->integer('temperatura');
             $table->integer('humedad');
             $table->integer('viento');
             $table->integer('lluvia');
